Split a video file by given time period/requirements.txt
 ffmpeg-python